% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{srtm_build_cutpoint_plot}
\alias{srtm_build_cutpoint_plot}
\title{Build a cut-point plot for inclusion in an SRTM report}
\usage{
srtm_build_cutpoint_plot(
  results,
  group_first = c("slope", "baseline"),
  plot_palette = "simple",
  debug = FALSE
)
}
\description{
Build a cut-point plot for inclusion in an SRTM report
}
\keyword{internal}
