% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slope_threshold.R
\name{srtm_getSlopeThresholds}
\alias{srtm_getSlopeThresholds}
\title{Compute slope thresholds for trajectory typing}
\usage{
srtm_getSlopeThresholds(scale_range, dt01, steep_frac = 0.25, flat_frac = 0.05)
}
\arguments{
\item{scale_range}{Numeric > 0. Approximate scale range of the outcome
(e.g., max(y*) - min(y*)).}

\item{dt01}{Numeric > 0. Time between y0 and y1.}

\item{steep_frac}{Numeric > 0. Fraction of scale_range per time unit
used to define "steep" slopes.}

\item{flat_frac}{Numeric > 0. Fraction of scale_range per time unit
used to define the central "flat" band.}
}
\description{
Compute slope thresholds for trajectory typing
}
\keyword{internal}
