% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTM_synth_data.R
\docType{data}
\name{SRTM_synth_data_100}
\alias{SRTM_synth_data_100}
\title{Example synthetic SRTM dataset with 100 cases (seed=391680921)}
\format{
A tibble with columns:
\describe{
\item{ID}{Individual ID}
\item{y0}{Baseline score}
\item{y1}{Score at time 1}
\item{y2}{Score at time 2}
}
}
\usage{
data("SRTM_synth_data")
}
\description{
Example synthetic SRTM dataset with 100 cases (seed=391680921)
}
\keyword{datasets}
