% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_element_helpers.R
\name{srtm_wrap_title}
\alias{srtm_wrap_title}
\title{Wrap a ggplot title safely}
\usage{
srtm_wrap_title(p, width = 30, size = 10)
}
\description{
Wrap a ggplot title safely
}
\keyword{internal}
