% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimePeriod.R
\name{getTimePeriod}
\alias{getTimePeriod}
\title{Internal helper to obtain a valid time period}
\usage{
getTimePeriod(
  time_period = NULL,
  interactive = TRUE,
  label = "between two time points"
)
}
\description{
Internal helper to obtain a valid time period
}
\keyword{internal}
