% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_type.R
\name{srtm_compute_trajType}
\alias{srtm_compute_trajType}
\title{Compute semantic trajectory types from slope groups}
\usage{
srtm_compute_trajType(
  data,
  slope_col = "m01",
  group_col = "trajGroup",
  label_scheme = c("arrows", "signs", "text"),
  dt01 = NULL,
  steep_const = 0.25,
  flat_const = 0.05
)
}
\description{
Compute semantic trajectory types from slope groups
}
\keyword{internal}
