% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-semantic_labels.R
\name{srtm_trajectory_label_schemes}
\alias{srtm_trajectory_label_schemes}
\title{Trajectory label schemes for SRTM}
\usage{
srtm_trajectory_label_schemes()
}
\description{
Trajectory label schemes for SRTM
}
\keyword{internal}
